👌 IMPROVE: Move test service inside services folder
