+++ client/src/app/pages/home/home.component.ts
@@ -4,8 +4,8 @@ import { Auth, User, getIdToken } from '@angular/fire/auth';
 import { FormGroup, FormsModule } from '@angular/forms';
 import { Router } from '@angular/router';
 import { UserProfile, UserRole } from '../../../../../shared/src/types/user';
-import { TestService } from '../../service/test.service';
 import { AuthenticationService } from '../../services/authentication.service';
+import { TestService } from '../../services/test.service';
 import { BusinessViewComponent } from '../../views/home/business-view/business-view.component';
 import { LandingViewComponent } from '../../views/home/landing-view/landing-view.component';
 import { UserViewComponent } from '../../views/home/user-view/user-view.component';
